requirements.txt
python-telegram-bot[webhooks]==20.7
requests
pydub
flask
gtts
python-dotenv
